+++ src/plugin.hpp
@@ -12,6 +12,7 @@ extern Plugin* pluginInstance;
 
 // Declare each Model, defined in each module source file
 
+extern Model* modelADREnvelope;
 extern Model* modelArm3XY;
 extern Model* modelAutoScale4;
 extern Model* modelBernoulliSwitch;